Completed the logic of 5 round Rock, Paper, Scissors
Completed the bare minimum of the assignment to play Rock, Paper and Scissors until either the player or the computer wins. Quite stoked with the results but as usual there are some clean codes and optimization to work on. Pretty much working but defintely has room for improvement. Will clean thee codes up soon!

[Ticket: 1 - Date 30/5/2021]